requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
python-jose[cryptography]>=3.3
bcrypt>=4.0
pydantic-settings>=2.0
python-multipart>=0.0.9
slowapi>=0.1.9
httpx>=0.27
pytest>=8.0
